src/main/java/com/anton/day3/validator/parser/NumberParser.java
package com.anton.day3.validator.parser;

public class NumberParser {
    public boolean isValidInt(String number) {
        try {
            Integer.parseInt(number);
            return true;
        } catch (NumberFormatException e) {
            return false;
        }
    }

    public boolean isValidDouble(String number) {
        if(number == null) {
            return false;
        }
        try {
            Double.parseDouble(number);
            return true;
        } catch (NumberFormatException e) {
            return false;
        }
    }
}
